Move WorkingCopyDirty and WorkingCopyTreeMismatch...
into kart/working_copy.py - they need not be table-specific.
